requirements.txt
flask~=3.0
pyjwt~=2.10
bcrypt~=4.2
python-dotenv~=1.0
requests~=2.32
